backend/requirements.txt
Flask==3.0.0
Flask-CORS==5.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.5
Flask-SQLAlchemy
pytz==2023.3
Werkzeug==3.0.1